Move to singleton, improve render perf
